requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==4.0.2
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
